# resources/swagger/georchestra-signalement.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Georchestra - Plugin Signalement API",
		"contact": {
			"email": "guillaume.morin@open-groupe.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "",
	"basePath": "/signalement",
...
	"schemes": [
		"https",
		"http"
	],
	"paths": {
...
		"/styles": {
			"get": {
				"tags": [
					"styles"
				],
				"summary": "Recherche de la liste des style.",
				"description": "Récupération de la liste des styles disponibles qui respectent les filtres.",
				"operationId": "searchStyles",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer",
						"default": 0
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer",
						"default": 10
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string",
						"default": ""
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/StylePageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"put": {
				"tags": [
					"styles"
				],
				"summary": "Mise à jour d'un style.",
				"description": "Mise à jour d' un style tel que stipulée dans le body de la requête, en fonction de l'id du style",
				"operationId": "updateStyle",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "style",
						"description": "Le nouveau stylé à appliquer",
						"required": true,
						"schema": {
							"$ref": "#/definitions/StyleContainer"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/StyleContainer"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},

			"post": {
			"tags": [
				"styles"
			],
			"summary": "Création d'un style",
			"description": "",
			"operationId": "createStyle",
			"produces": [
				"application/json"
			],
			"parameters": [
				{
					"in": "body",
					"name": "style",
					"description": "Description du style",
					"required": true,
					"schema": {
						"$ref": "#/definitions/StyleContainer"
					}
				}
			],
			"responses": {
				"200": {
					"description": "OK",
					"schema": {
						"$ref": "#/definitions/StyleContainer"
					}
				},
				"400": {
					"description": "Bad input parameter"
				},
				"500": {
					"description": "Internal server error",
					"schema": {
						"$ref": "#/definitions/ApiError"
					}
				}
			}
		}

		},
		"/styles/{id}": {
			"delete": {
				"tags": [
					"styles"
				],
				"summary": "Suppression d'un style.",
				"description": "Suppression d'un style en fonction de son id",
				"operationId": "deleteStyle",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "id du style à supprimer",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/styles/process": {
			"post": {
				"tags": [
					"styles"
				],
				"summary": "Création d'une liaison (avec les tasks de workflow) associé à ce style",
				"description": "",
				"operationId": "createProcessStyling",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "processStyle",
						"description": "Description de la liaison du style avec une task",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ProcessStyling"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ProcessStyling"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/styles/process/{id}": {
			"get": {
				"tags": [
					"styles"
				],
				"summary": "Récupération liaisons (avec les tasks de workflow) associé à ce style",
				"description": "",
				"operationId": "getStyleProcessById",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "id du style dont les liaisons sont à récupérer",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/ProcessStyling"
							}
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"styles"
				],
				"summary": "Suppression d'une liaison (avec les tasks de workflow) associé à ce style.",
				"description": "Suppression d'un ProcessStyling en fonction de son id",
				"operationId": "deleteProcessStyling",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "id du ProcessStyling à supprimer",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
		"/task/draft": {
			"post": {
				"tags": [
					"tasks"
				],
				"summary": "Création d'un signalement brouillon",
				"description": "",
				"operationId": "createDraft",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "reportingDescription",
						"description": "Description du signalement",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ReportingDescription"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/task/cancel/{uuid}": {
			"delete": {
				"tags": [
					"tasks"
				],
				"summary": "Abandon d'un signalement brouillon",
				"description": "",
				"operationId": "cancelDraft",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "uuid",
						"description": "Identifiant du signalement",
						"required": true,
						"type": "string",
						"format": "uuid"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
		"/task/start": {
			"post": {
				"tags": [
					"tasks"
				],
				"summary": "Démarrage du workflow de signalement",
				"description": "",
				"operationId": "startTask",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "task",
						"description": "Description du signalement",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Task"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/task/claim/{taskId}": {
			"put": {
				"tags": [
					"tasks"
				],
				"summary": "Demande d'affectation",
				"description": "",
				"operationId": "claimTask",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"description": "Identifiant de la tâche associée",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/task/{taskId}": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "Récupérer une tâche d'un signalement",
				"description": "",
				"operationId": "getTask",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"description": "Identifiant de la tâche associée",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/task/update": {
			"put": {
				"tags": [
					"tasks"
				],
				"summary": "Mise à jour de la tâche du signalement",
				"description": "",
				"operationId": "updateTask",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "task",
						"description": "Description du signalement",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Task"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Task"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
		"/task/search": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "Liste les tâches du user courant",
				"description": "Retourne la liste des tâches de l'utilisateur connecté",
				"operationId": "searchTasks",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "contextName",
						"description": "Nom du contexte pour filtrage. Pas de filtrage si null",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "contextType",
						"description": "Type du contexte pour filtrage",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "geographicType",
						"description": "Type géographique pour filtrage",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "asAdmin",
						"description": "En tant qu'administrateur. Faux par défaut",
						"required": false,
						"type": "boolean"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Task"
							}
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/task/search/geojson": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "Liste les tâches du user courant au format GeoJSon",
				"description": "Retourne la liste des tâches de l'utilisateur connecté au format GeoJSon",
				"operationId": "searchGeoJSonTasks",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "contextName",
						"description": "Nom du contexte pour filtrage",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "contextType",
						"description": "Type du contexte pour filtrage",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "geographicType",
						"description": "Type géographique pour filtrage",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "asAdmin",
						"description": "En tant qu'administrateur. Faux par défaut",
						"required": false,
						"type": "boolean"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/FeatureCollection"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/task/geojson/properties": {
			"get": {
				"tags": [
					"tasks"
				],
				"summary": "Liste les description des propriétés disponibles dans le GeoJson",
				"description": "Retourne la description des propriétés disponibles dans une feature GeoJson. Si le contexte géographique n'est pas précisé, aucune propriété geometry n'est présente",
				"operationId": "getGeoJSonTaskProperties",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "contextName",
						"description": "Nom du contexte géographique",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/FeatureTypeDescription"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
	},
	"definitions": {
		"ApiError": {
			"type": "object",
			"properties": {
				"code": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
...
		"PointG": {
			"type": "object",
			"properties": {
				"x": {
					"type": "string"
				},
				"y": {
					"type": "string"
				}
			}
		},
		"GeographicType": {
			"type": "string",
			"description": "Type géographic",
			"enum": [
				"POINT",
				"LINE",
				"POLYGON"
			]
		},
...
		"ContextType": {
			"type": "string",
			"description": "Type du context du signalement",
			"enum": [
				"THEMA",
				"LAYER"
			]
		},
		"FieldType": {
			"type": "string",
			"description": "Type des champs",
			"enum": [
				"LONG",
				"DOUBLE",
				"STRING",
				"BOOLEAN",
				"DATE",
				"LIST"
			]
		},
		"ValidatorType": {
			"type": "string",
			"enum": [
				"REQUIRED",
				"MAXLENGTH",
				"POSITIVE",
				"NEGATIVE",
				"REGEXP"
			]
		},
		"Validator": {
			"type": "object",
			"properties": {
				"type": {
					"$ref": "#/definitions/ValidatorType"
				},
				"attribute": {
					"type": "string"
				}
			}
		},
		"FieldDefinition": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"help": {
					"type": "string"
				},
				"type": {
					"type": "object",
					"$ref": "#/definitions/FieldType"
				},
				"readOnly": {
					"type": "boolean"
				},
				"required": {
					"type": "boolean"
				},
				"multiple": {
					"type": "boolean"
				},
				"extendedType": {
					"type": "string"
				},
				"validators": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Validator"
					}
				}
			}
		},
		"Field": {
			"type": "object",
			"properties": {
				"values": {
					"type": "array",
					"items": {
						"type": "string"
					}
				},
				"definition": {
					"type": "object",
					"$ref": "#/definitions/FieldDefinition"
				}
			}
		},
		"Section": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"readOnly": {
					"type": "boolean"
				},
				"fields": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/Field"
					}
				}
			}
		},
		"Form": {
			"type": "object",
			"properties": {
				"sections": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/Section"
					}
				}
			}
		},
		"ContextDescription": {
			"type": "object",
			"properties": {
				"contextType": {
					"type": "object",
					"$ref": "#/definitions/ContextType"
				},
				"geographicType": {
					"type": "object",
					"$ref": "#/definitions/GeographicType"
				},
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"processDefinitionKey": {
					"type": "string"
				},
				"revision": {
					"type": "integer"
				}
			}
		},
...
		"Status": {
			"type": "string",
			"description": "Etat d'un signalement",
			"enum": [
				"DRAFT",
				"PENDING",
				"CANCELLED",
				"COMPLETED",
				"DELETED"
			]
		},
		"Attachment": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"name": {
					"type": "string"
				},
				"mimeType": {
					"type": "string"
				}
			}
		},
...
		"ReportingDescription": {
			"type": "object",
			"properties": {
				"functionalId": {
					"type": "integer",
					"format": "int64"
				},
				"uuid": {
					"type": "string",
					"format": "uuid"
				},
				"status": {
					"type": "object",
					"$ref": "#/definitions/Status"
				},
				"functionalStatus": {
					"type": "string"
				},
				"geographicType": {
					"type": "object",
					"$ref": "#/definitions/GeographicType"
				},
				"contextDescription": {
					"type": "object",
					"$ref": "#/definitions/ContextDescription"
				},
				"initiator": {
					"type": "string"
				},
				"creationDate": {
					"type": "string",
					"format": "date-time"
				},
				"updatedDate": {
					"type": "string",
					"format": "date-time"
				},
				"description": {
					"type": "string"
				},
				"assignee": {
					"type": "string"
				},
				"localisation": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/PointG"
					}
				},
				"attachments": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Attachment"
					}
				},
				"datas": {
					"type": "object"
				}
			}
		},
		"Action": {
			"type": "object",
			"properties": {
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
...
		"Task": {
			"type": "object",
			"properties": {
				"id": {
					"type": "string"
				},
				"functionalId": {
					"type": "integer",
					"format": "int64"
				},
				"status": {
					"type": "object",
					"$ref": "#/definitions/Status"
				},
				"functionalStatus": {
					"type": "string"
				},
				"initiator": {
					"type": "string"
				},
				"creationDate": {
					"type": "string",
					"format": "date-time"
				},
				"updatedDate": {
					"type": "string",
					"format": "date-time"
				},
				"assignee": {
					"type": "string"
				},
				"actions": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Action"
					}
				},
				"form": {
					"type": "object",
					"$ref": "#/definitions/Form"
				},
				"asset": {
					"type": "object",
					"$ref": "#/definitions/ReportingDescription"
				}
			}
		},
		"GeometryType": {
			"type": "string",
			"description": "the geometry type",
			"enum": [
				"Point",
				"LineString",
				"Polygon",
				"MultiPoint",
				"MultiLineString",
				"MultiPolygon"
			]
		},
		"Geometry": {
			"type": "object",
			"description": "GeoJSon geometry",
			"discriminator": "type",
			"properties": {
				"type": {
					"type": "object",
					"$ref": "#/definitions/GeometryType"
				}
			}
		},
...
		"Style": {
			"type": "object",
			"description": "GeoJSon feature style for frontend",
			"properties": {
				"color": {
					"type": "string"
				},
				"fillColor": {
					"type": "string"
				},
				"fillOpacity": {
					"type": "number",
					"format": "double"
				},
				"opacity": {
					"type": "number",
					"format": "double"
				},
				"iconColor": {
					"type": "string"
				},
				"iconGlyph": {
					"type": "string"
				},
				"iconShape": {
					"type": "string"
				},
				"weight": {
					"type": "number",
					"format": "double"
				},
				"geometry": {
					"type": "string"
				},
				"type": {
					"type": "string"
				},
				"iconAnchor": {
					"type": "array",
					"items":{
						"type": "number",
						"format": "double"
					}
				},
				"dashArray": {
					"type": "array",
					"items": {
						"type": "number",
						"format": "double"
					}
				},
				"filtering": {
					"type": "boolean"
				}
			}
		},
		"StyleContainer": {
			"type": "object",
			"description": "GeoJSon feature style which make easier to go from database to backend",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"type": {
					"type": "string",
					"$ref": "#/definitions/GeographicType"
				},
				"style": {
					"type": "object",
					"$ref": "#/definitions/Style"
				},
				"name": {
					"type": "string"
				}
			}
		},
		"ProcessStyling": {
			"type": "object",
			"description": "Link between a user task and GeoJSon feature style",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"processDefinitionId": {
					"type": "string"
				},
				"revision": {
					"type": "integer"
				},
				"userTaskId": {
					"type": "string"
				},
				"stylingId": {
					"type": "integer",
					"format": "int64"
				}
			}
		},
		"StylePageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/StyleContainer"
					}
				}
			}
		},
...
		"Feature": {
			"type": "object",
			"description": "GeoJSon feature",
			"properties": {
				"type": {
					"type": "string",
					"enum": [
						"Feature"
					]
				},
				"id": {
					"type": "string",
					"format": "uuid"
				},
				"geometry": {
					"$ref": "#/definitions/Geometry"
				},
				"style": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Style"
					}
				},
				"properties": {
					"type": "object"
				}
			}
		},
		"FeatureCollection": {
			"type": "object",
			"description": "GeoJSon feature collection",
			"properties": {
				"type": {
					"type": "string",
					"enum": [
						"FeatureCollection"
					]
				},
				"features": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Feature"
					}
				},
				"style": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Style"
					}
				}
			}
		},
		"FeatureProperty": {
			"type": "object",
			"description": "GeoJSon feature property description",
			"properties": {
				"name": {
					"type": "string"
				},
				"maxOccurs": {
					"type": "integer",
					"format": "int32"
				},
				"minOccurs": {
					"type": "integer",
					"format": "int32"
				},
				"nillable": {
					"type": "boolean"
				},
				"type": {
					"type": "string"
				},
				"localType": {
					"type": "string"
				}
			}
		},
		"FeatureType": {
			"type": "object",
			"description": "GeoJSon feature type",
			"properties": {
				"typeName": {
					"type": "string"
				},
				"properties": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/FeatureProperty"
					}
				}
			}
		},
		"FeatureTypeDescription": {
			"type": "object",
			"description": "GeoJSon feature type description",
			"properties": {
				"elementFormDefault": {
					"type": "string"
				},
				"targetNamespace": {
					"type": "string"
				},
				"targetPrefix": {
					"type": "string"
				},
				"featureTypes": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/FeatureType"
					}
				}
			}
		},
...
	}
}